% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{act_partner}
\alias{act_partner}
\title{ACT partner drug resistance}
\format{
A tibble with 10334 rows and 22 columns:
}
\source{
https://www.wwarn.org/tracking-resistance/act-partner-drug-molecular-surveyor
}
\usage{
act_partner
}
\description{
Data on molecular markers for resistance to ACT partner drugs namely
piperaquine (PPQ), amodiaquine (AQ), lumefantrine (LUM) and mefloquine (MQ)
found in Plasmodium falciparum malaria parasite pfmdr1, pfcrt, and pfpm2
genes
}
\examples{
act_partner

}
\keyword{datasets}
