% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicine_quality}
\alias{medicine_quality}
\title{Antimalarial medicine quality}
\format{
A tibble with 1553 rows and 37 columns:
}
\source{
https://www.wwarn.org/about-us/medicine-quality/mq-scientific-literature-surveyor-antimalarial-quality-data
}
\usage{
medicine_quality
}
\description{
Substandard and falsified (SF) antimalarial medicines pose an immediate
danger to vulnerable populations and could increase the threat of malaria
drug resistance emerging and spreading. The major challenge in preventing
the spread and distribution of poor quality medical products is the lack of
accessible and reliable information on their prevalence.
}
\examples{
medicine_quality

}
\keyword{datasets}
