% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sulphadoxine}
\alias{sulphadoxine}
\title{Sulphadoxine pyrimethamine (SP) resistance}
\format{
A tibble with 7855 rows and 22 columns:
}
\source{
https://www.wwarn.org/tracking-resistance/sp-molecular-surveyor
}
\usage{
sulphadoxine
}
\description{
Sulphadoxine pyrimethamine (SP) is the recommended intermittent preventive
treatment (IPTp) against malaria infection for pregnant women and
SP + Amodiaquine is the combination used for seasonal malaria
chemoprophylaxis (SMC).  Because these two preventive treatments are now
being widely deployed, there is renewed interest in the level of resistance
to these antimalarials. This dataset provides information on the prevalence
of molecular markers associated with resistance to SP found in
\emph{Plasmodium falciparum} malaria parasite by year and location.
}
\examples{
sulphadoxine

}
\keyword{datasets}
