% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vivax}
\alias{vivax}
\title{Plasmodium vivax clinical trials}
\format{
A tibble with 16326 rows and 44 columns:
}
\source{
https://www.wwarn.org/tracking-resistance/vivax-surveyor
}
\usage{
vivax
}
\description{
Data on prevalence of *Plasmodium vivax* antimalarial clinical trials across
the world. The data are derived from a recent review of *Plasmodium vivax*
antimalarial clinical trials.
}
\examples{
vivax

}
\keyword{datasets}
