% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwarn.R
\docType{package}
\name{wwarnsurveyr}
\alias{wwarnsurveyr}
\alias{wwarnsurveyr-package}
\title{WWARN Antimalarial Resistance Surveyors in R}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
