% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artemisinin}
\alias{artemisinin}
\title{Artemisinin resistance based on molecular marker in the Kelch 13 gene}
\format{
A tibble with 2434 rows and 16 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{continent} \tab Name of continent \cr
   \code{val} \tab Effect on clearance \cr
   \code{country} \tab Name of country \cr
   \code{year} \tab Year \cr
   \code{tested} \tab Number of cases tested \cr
   \code{prev} \tab Prevalence of Kelch 13 gene molecular marker \cr
   \code{lon} \tab Longitude coordinate for site location \cr
   \code{title} \tab Title of paper from which data was derived \cr
   \code{pubMedId} \tab PubMed identifier \cr
   \code{sid} \tab Site identifier \cr
   \code{mutation} \tab Type of mutation \cr
   \code{site} \tab Name of site \cr
   \code{estLoc} \tab Location \cr
   \code{present} \tab Number of cases where mutation was present \cr
   \code{lat} \tab Latitude coordinate for site location \cr
   \code{authors} \tab Name of authors of paper from which data was derived \cr
}
}
\source{
https://www.wwarn.org/tracking-resistance/artemisinin-molecular-surveyor
}
\usage{
artemisinin
}
\description{
The discovery of molecular markers of resistance in the Kelch 13 gene gives
us the unusual opportunity to monitor the emergence and spread of resistance
nearly in real time. Assessing the prevalence of these molecular markers in
other regions can provide an early warning system to trigger rapid responses
to develop control strategies.
}
\details{
The Artemisinin Molecular Surveyor is an interactive map that summarises the
prevalence of these molecular markers in the propeller region of the Kelch 13
gene of the malaria parasite by location and year, providing a clear and
standardised visualisation of the evidence to date to inform key
international, regional and national monitoring strategies.
}
\examples{
artemisinin

}
\keyword{datasets}
