% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwarnsurveyr.R
\docType{package}
\name{wwarnsurveyr}
\alias{wwarnsurveyr}
\alias{wwarnsurveyr-package}
\title{WWARN Antimalarial Resistance Surveyors in R}
\description{
The WorldWide Antimalarial Resistance Network (WWARN)'s mission is
to generate innovative tools and reliable evidence to inform the malaria
community on the factors affecting the efficacy of antimalarial medicines.
WWARN works with collaborators to optimise the efficacy of antimalarial
medicines and treatment regimens, especially for vulnerable groups including
pregnant women, infants and malnourished children, and provides evidence to
inform the development of new antimalarial drugs. WWARN also engages and
works with other disease research communities to replicate this model for
other tropical and neglected diseases. WWARN's interactive maps summarise
data from published studies and data stored in the WWARN data centre.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OxfordIHTM/wwarnsurveyr}
  \item Report bugs at \url{https://github.com/OxfordIHTM/wwarnsurveyr/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
